Small optimizations and fixes
- Change `prototype->arity/upvalue_count` types to `uint8_t`
- Update parameter max check in `_compiler_function_expr`
- Fix `OP_PUSH_ARRAY` and `OP_PUSH_MAP` in `wi_disasm.c` by adding a separate `_short_instr` function
- Reorder `wi_value_hash`, which does increase speed
- Bump version to `v4.1.2-beta`
- Update `test/benchmark/README.md` accordingly

+++ src/core/wi_box.c
@@ -158,7 +158,7 @@ wi_closure_t*
 wi_new_closure(wi_gc_t* gc, wi_prototype_t* prototype, wi_table_t* globals) {
     wi_upvalue_t** upvalues = WI_GC_ALLOC(gc, wi_upvalue_t*, prototype->upvalue_count);
 
-    for (int i = 0; i < prototype->upvalue_count; i++) {
+    for (uint8_t i = 0; i < prototype->upvalue_count; i++) {
         upvalues[i] = NULL;
     }
